+++ ejbModule/service/ResearchGroupService.java
@@ -10,11 +10,11 @@ public interface ResearchGroupService {
 
 	public boolean addResearchGroup(String name,int deptId, String description, String url, String photoUrl) throws YearbookException;
 	
-	public ResearchGroup getResearchGroup(int groupId) throws YearbookException;
+	public ResearchGroup getResearchGroup(long groupId) throws YearbookException;
 	
-	public boolean updateResearchGroup(int groupId, String name, int deptId, String description, String url, String photoUrl) throws YearbookException;
+	public boolean updateResearchGroup(long groupId, String name, int deptId, String description, String url, String photoUrl) throws YearbookException;
 	
-	public boolean deleteReseachGroup(int groupId) throws YearbookException;
+	public boolean deleteReseachGroup(long groupId) throws YearbookException;
 
 	public List<ResearchGroup> getAllResearchGroups() throws YearbookException;
